Move stateless to base; switch to EmptyState
